Naming convention fix.
