+++ vector/src/main/java/im/vector/app/features/location/LocationSharingSettingsNavigator.kt
@@ -16,21 +16,8 @@
 
 package im.vector.app.features.location
 
-import android.app.Activity
-import im.vector.app.core.utils.openAppSettingsPage
-
-class LocationSharingNavigator constructor(val activity: Activity?) : ILocationSharingNavigator {
-
-    override var goingToAppSettings: Boolean = false
-
-    override fun quit() {
-        activity?.finish()
-    }
-
-    override fun goToAppSettings() {
-        activity?.let {
-            goingToAppSettings = true
-            openAppSettingsPage(it)
-        }
-    }
+interface LocationSharingSettingsNavigator {
+    var goingToAppSettings: Boolean
+    fun quit()
+    fun goToAppSettings()
 }
